+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 ## v1.10.0 (December 29, 2014)
 * *[v1.10.0.rc1]*
+  * [FEATURE]: Added support for optional controls via card data
   * [FEATURE]: Added GA timing event for how long it takes c6ui to
     download
   * [FEATURE]: Add GA timing event for how long it takes a video to